pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "discord-py>=2.6.2",
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "lavalink>=5.9.0",
    "psycopg2-binary>=2.9.10",
    "pynacl>=1.6.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "twilio>=9.7.2",
    "werkzeug>=3.1.3",
]
